+++ miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/rest/ContainerRestController.java
@@ -39,6 +39,7 @@ import javax.ws.rs.core.Response.Status;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
@@ -89,6 +90,9 @@ public class ContainerRestController extends RestController {
   private SequencingContainerModelService containerModelService;
   @Autowired
   private IndexChecker indexChecker;
+
+  @Value("${miso.detailed.sample.enabled}")
+  private Boolean detailedSample;
 
   @Autowired
   private AdvancedSearchParser advancedSearchParser;
@@ -156,7 +160,7 @@ public class ContainerRestController extends RestController {
     Stream<Partition> input = request.getIds().stream()
         .map(WhineyFunction.rethrow(containerService::get))
         .flatMap(container -> container.getPartitions().stream());
-    return MisoWebUtils.generateSpreadsheet(request, input, PartitionSpreadsheets::valueOf, response);
+    return MisoWebUtils.generateSpreadsheet(request, input, detailedSample, PartitionSpreadsheets::valueOf, response);
   }
 
   private static class SerialNumberValidationDto {